Add offset= to backup info durable string.
