+++ src/btree/bt_import.c
@@ -96,7 +96,7 @@ __wt_import(WT_SESSION_IMPL *session, const char *uri)
      */
     filecfg[1] = a->data;
     filecfg[2] = checkpoint_list;
-    filecfg[3] = "checkpiont_backup_info=";
+    filecfg[3] = "checkpoint_backup_info=";
     filecfg[4] = "checkpoint_lsn=";
     WT_WITH_SCHEMA_LOCK(session,
       ret = __wt_snprintf(fileid, sizeof(fileid), "id=%" PRIu32, ++S2C(session)->next_file_id));
