+++ src/include/wt_internal.h
@@ -370,6 +370,7 @@ typedef uint64_t wt_timestamp_t;
 #include "posix.h"
 #endif
 
+#include "meta.h" /* required by block.h */
 #include "misc.h"
 #include "mutex.h"
 
@@ -392,7 +393,6 @@ typedef uint64_t wt_timestamp_t;
 #include "error.h"
 #include "log.h"
 #include "lsm.h"
-#include "meta.h"
 #include "optrack.h"
 #include "os.h"
 #include "reconcile.h"
